+++ src/Api/Extensions/HttpRequestExtensions.cs
@@ -7,17 +7,7 @@ public static class HttpRequestExtensions
 {
     public static string? GetApiSecret(this HttpRequest req) => req.Headers.GetApiSecret();
 
-    public static string? GetApiSecret(this IHeaderDictionary headerDictionary) => headerDictionary.GetHeaderValue("ApiSecret");
-
     public static string? GetPublicApiKey(this HttpRequest req) => req.Headers.GetPublicApiKey();
-
-    public static string? GetPublicApiKey(this IHeaderDictionary headerDictionary) => headerDictionary.GetHeaderValue("ApiKey");
-
-    private static string? GetHeaderValue(this IHeaderDictionary headerDictionary, string key)
-    {
-        headerDictionary.TryGetValue(key, out var value);
-        return value.SingleOrDefault();
-    }
 
     public static string? GetTenantNameFromKey(this HttpRequest request)
     {
